Splitting out Mask tests to a new file.
